% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_docx.R
\name{dml_docx}
\alias{dml_docx}
\title{A Microsoft Word DrawingML file}
\usage{
dml_docx(file = "Rplots.dml", width = 10, height = 8, bg = "white",
  fontname_serif = getOption("rvg_fonts")$fontname_serif,
  fontname_sans = getOption("rvg_fonts")$fontname_sans,
  fontname_mono = getOption("rvg_fonts")$fontname_mono,
  fontname_symbol = getOption("rvg_fonts")$fontname_symbol, pointsize = 12,
  init_id = 1, editable = TRUE)
}
\arguments{
\item{file}{the file where output will appear.}

\item{height, width}{Height and width in inches.}

\item{bg}{Default background color for the plot (defaults to "white").}

\item{fontname_serif, fontname_sans, fontname_mono, fontname_symbol}{font names for font faces}

\item{pointsize}{default point size.}

\item{init_id}{this id must will be used as the first unique id of graphical elements
to produce within a word document.}

\item{editable}{should vector graphics elements (points, text, etc.) be editable.}
}
\description{
A Microsoft Word DrawingML file
}
\examples{
dml_docx( file = tempfile() )
plot(1:11,(-5:5)^2, type='b', main="Simple Example")
dev.off()
}
\keyword{device}

