% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_docx.R
\name{docx_add_vgplot}
\alias{docx_add_vgplot}
\title{vml graph code}
\usage{
docx_add_vgplot(x, code, pos = "after", ...)
}
\arguments{
\item{x}{a docx device}

\item{code}{plot instructions}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{...}{arguments passed on to \code{\link{dml_docx}}.}
}
\description{
produces the vml of a graph
}
