% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveTextGrob.R
\name{drawDetails.interactiveTextGrob}
\alias{drawDetails.interactiveTextGrob}
\title{interactiveTextGrob drawing}
\usage{
\method{drawDetails}{interactiveTextGrob}(x, recording)
}
\arguments{
\item{x}{ A grid grob. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
draw an interactiveTextGrob
}

