% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsvg.R
\name{htmlDSVG}
\alias{htmlDSVG}
\title{Run plotting code and view svg in RStudio Viewer or web broswer.}
\usage{
htmlDSVG(code, ...)
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{dsvg}}.}
}
\description{
This is useful primarily for testing. Requires the \code{htmltools}
package.
}
\examples{
if (require("htmltools")) {
  htmlDSVG(plot(1:10))
  htmlDSVG(hist(rnorm(100)))
}
}

