% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pptx.R
\name{placeholder_add_vgplot}
\alias{placeholder_add_vgplot}
\title{pml graph code}
\usage{
placeholder_add_vgplot(x, code, id, index = 1, ...)
}
\arguments{
\item{x}{a pptx device}

\item{code}{\code{flextable} object}

\item{id}{placeholder id}

\item{index}{placeholder index. This is to be used when a placeholder id
is not unique in the current slide, e.g. two placeholders with id 'body'.}

\item{...}{arguments passed on to \code{\link{dml_pptx}}.}
}
\description{
produces the pml of a graph
}

