% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rvg-defunct}
\alias{rvg-defunct}
\alias{dml_docx}
\alias{body_add_vg}
\title{Defunct Functions in Package rvg}
\usage{
dml_docx(...)

body_add_vg(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package rvg
}
\details{
\code{dml_docx()} is removed and can be replaced by \code{devEMF::emf}.

\code{body_add_vg()} is removed and can be replaced by \code{officer::body_add_img}.
}
