% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rvg.R
\name{rvg}
\alias{rvg}
\title{Create SVG Graphic}
\usage{
rvg(rootname, width = 504, height = 504, ps = 12,
  fontname = "Helvetica", plot_id = 1)
}
\arguments{
\item{rootname}{root filename to create. For each plot, a file is produced.
The name of the file is the concatenation of \code{rootname}, \code{plot_id}
plot index (it will be incremented for each new plot) and the extention \code{.svg}.}

\item{width}{width in pixel}

\item{height}{height in pixel}

\item{ps}{point size}

\item{fontname}{font name}

\item{plot_id}{starting plot index. Default to 1.}
}
\description{
Create SVG Graphic
}
\examples{
rvg("myplot")
with( iris, plot( Sepal.Length, Petal.Length ) )
dev.off()

# an interactive example
rvg("myiplot")
    with( iris, plot( Sepal.Length, Petal.Length , type = "n" ) )

   sdata = split( iris, iris$Species )

   for(i in names( sdata ) ){
     tempdata = sdata[[i]]
     rvg_tracer_on()
     points(x = tempdata$Sepal.Length , y = tempdata$Petal.Length, pch = 16 )
     ids = rvg_tracer_off()

     sendTooltip(as.integer(ids), as.character(tempdata$Petal.Length) )
     sendClick(as.integer(ids), paste0("function(){alert('", tempdata$Species,"')}") )
    }
dev.off()
}

