% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rvg_tracer_off}
\alias{rvg_tracer_off}
\title{trace off id colection}
\usage{
rvg_tracer_off()
}
\value{
graphical elements id as integer values.
}
\description{
get collected id of an rvg device and
stop collecting.
}

