% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{sendClick}
\alias{sendClick}
\title{send click}
\usage{
sendClick(ids, clicks)
}
\arguments{
\item{ids}{integer vector of graphical elements identifiers (returned by
\code{\link{rvg_tracer_off}}).}

\item{clicks}{javascript functions to execute on click actions.}
}
\description{
add tooltips on graphical elements.
}

