% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_attr}
\alias{set_attr}
\title{set attributes to graphical elements}
\usage{
set_attr(ids, attribute, str)
}
\arguments{
\item{ids}{integer vector of graphical elements identifiers (returned by
\code{\link{rvg_tracer_off}}).}

\item{attribute}{name of the attribute to set.}

\item{str}{values to set for the attribute.}
}
\description{
set attributes with javascript instructions
to graphical elements.
}

