% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pptx.R
\name{write_pptx}
\alias{write_pptx}
\title{Microsoft PowerPoint Graphics Device}
\usage{
write_pptx(file, code, ...)
}
\arguments{
\item{file}{filename of the Microsoft PowerPoint document to produce. File
extension must be \code{.pptx}.}

\item{code}{Plotting code to execute}

\item{...}{arguments passed on to \code{\link{dml_pptx}}.}
}
\description{
A graphical device for Microsoft PowerPoint documents.
}
\examples{
\donttest{
write_pptx(file = "my_plot_1.pptx", code = plot(rnorm(10)))
write_pptx(file = "my_plot_2.pptx", code = barplot(1:7, col = 1:7))
}
}
\keyword{device}

