% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx_vg_at.R
\name{xlsx_vg}
\alias{xlsx_vg}
\title{add a plot output as vector graphics into an Excel object}
\usage{
xlsx_vg(x, sheet_id, code, left, top, width, height, ...)
}
\arguments{
\item{x}{an \code{rxlsx} object produced by \code{officer::read_xlsx}}

\item{sheet_id}{sheet id}

\item{code}{plot instructions}

\item{left, top}{left and top origin of the plot on the slide in inches.}

\item{height, width}{Height and width in inches.}

\item{...}{arguments passed on to \code{\link{dml_xlsx}}.}
}
\description{
produces a vector graphics output from R plot instructions
and add the result in an Excel sheet.
by \code{\link[officer]{read_xlsx}}.
}
