% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_vg.R
\name{body_add_vg}
\alias{body_add_vg}
\title{vml graph code}
\usage{
body_add_vg(x, code, pos = "after", ...)
}
\arguments{
\item{x}{a docx device}

\item{code}{plot instructions}

\item{pos}{where to add the new element relative to the cursor,
one of "after", "before", "on".}

\item{...}{arguments passed on to \code{\link{dml_docx}}.}
}
\description{
produces the vml of a graph
}
