% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_pptx.R
\name{dml_pptx}
\alias{dml_pptx}
\title{A Microsoft PowerPoint DrawingML file}
\usage{
dml_pptx(file = "Rplots.dml", width = 10, height = 8, offx = 1,
  offy = 1, bg = "white",
  fontname_serif = getOption("rvg_fonts")$fontname_serif,
  fontname_sans = getOption("rvg_fonts")$fontname_sans,
  fontname_mono = getOption("rvg_fonts")$fontname_mono,
  fontname_symbol = getOption("rvg_fonts")$fontname_symbol, pointsize = 12,
  editable = TRUE)
}
\arguments{
\item{file}{the file where output will appear.}

\item{height, width}{Height and width in inches.}

\item{offx, offy}{top and left origin of the plot}

\item{bg}{Default background color for the plot (defaults to "white").}

\item{fontname_serif, fontname_sans, fontname_mono, fontname_symbol}{font names for font faces}

\item{pointsize}{default point size.}

\item{editable}{should vector graphics elements (points, text, etc.) be editable.}
}
\description{
A Microsoft PowerPoint DrawingML file
}
\examples{
dml_pptx( file = tempfile() )
plot(1:11,(-5:5)^2, type='b', main="Simple Example")
dev.off()
}
\keyword{device}

