% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsvg.R
\name{dsvg}
\alias{dsvg}
\title{A SVG Graphics Driver}
\usage{
dsvg(file = "Rplots.svg", width = 10, height = 8, bg = "white",
  pointsize = 12, standalone = TRUE, canvas_id = 1)
}
\arguments{
\item{file}{the file where output will appear.}

\item{height,width}{Height and width in inches.}

\item{bg}{Default background color for the plot (defaults to "white").}

\item{pointsize}{default point size.}

\item{standalone}{Produce a stand alone svg file? If \code{FALSE}, omits
xml header and default namespace.}

\item{canvas_id}{svg id within HTML page.}
}
\description{
This function produces graphics suitable the current w3 svg XML standard.
It currently does not have any font metric information, so the use of
\code{\link{plotmath}} is not supported. The driver output is currently NOT
specifying a DOCTYPE DTD
}
\examples{
dsvg()
plot(1:11,(-5:5)^2, type='b', main="Simple Example")
dev.off()
}
\seealso{
\code{\link{pictex}}, \code{\link{postscript}}, \code{\link{Devices}}
}
\keyword{device}

