% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pptx.R
\name{ph_with_vg}
\alias{ph_with_vg}
\title{pml graph code}
\usage{
ph_with_vg(x, code, type, index = 1, ...)
}
\arguments{
\item{x}{a pptx device}

\item{code}{plot instructions}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}

\item{...}{arguments passed on to \code{\link{dml_pptx}}.}
}
\description{
produces the pml of a graph
}
