% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_device.R
\name{pptx_device}
\alias{pptx_device}
\title{Microsoft PowerPoint Graphics Device}
\usage{
pptx_device(file, code, size = c(width = 10, height = 7.5), ...)
}
\arguments{
\item{file}{filename of the Microsoft PowerPoint document to produce. File
extension must be \code{.pptx}.}

\item{code}{Plotting code to execute}

\item{size}{slide size in inches.}

\item{...}{arguments for \code{fun} (passed on to \code{\link{dml_pptx}}.)}
}
\description{
A graphical device for Microsoft PowerPoint documents.
}
\examples{
pptx_device(file = "my_plot_1.pptx", code = plot(rnorm(10)))
pptx_device(file = "my_plot_2.pptx", code = barplot(1:7, col = 1:7))
}
\keyword{device}

