% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rvg-defunct}
\alias{rvg-defunct}
\alias{ph_with_vg}
\alias{ph_with_vg_at}
\title{Defunct Functions in Package rvg}
\usage{
ph_with_vg(...)

ph_with_vg_at(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package rvg
}
\details{
\code{ph_with_vg()} is replaced by \code{officer::ph_with}.

\code{ph_with_vg_at()} is replaced by \code{officer::ph_with}.
}
