% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{rvg-defunct}
\alias{rvg-defunct}
\alias{body_add_vg}
\title{Defunct Functions in Package rvg}
\usage{
body_add_vg(...)
}
\description{
Defunct Functions in Package rvg
}
\details{
\code{body_add_vg()} is removed and can be replaced by \code{officer::body_add_img}.
}
