% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsvg.R
\name{rvg}
\alias{rvg}
\title{Create SVG Graphic}
\usage{
rvg(file, width = 504, height = 504, ps = 12, fontname = "Helvetica",
  canvas_id = 0)
}
\arguments{
\item{file}{filename to create}

\item{width}{width in pixel}

\item{height}{height in pixel}

\item{ps}{point size}

\item{fontname}{font name}

\item{canvas_id}{canvas id}
}
\description{
Create SVG Graphic
}

