% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{send_click}
\alias{send_click}
\title{send click}
\usage{
send_click(ids, clicks)
}
\arguments{
\item{ids}{integer vector of graphical elements identifiers (returned by
\code{\link{rvg_tracer_off}}).}

\item{clicks}{javascript functions to execute on click actions.}
}
\description{
add tooltips on graphical elements.
}

