% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{sendTooltip}
\alias{sendTooltip}
\title{send tooltip}
\usage{
sendTooltip(ids, tooltips)
}
\arguments{
\item{ids}{integer vector of graphical elements identifiers (returned by
\code{\link{rvg_tracer_off}}).}

\item{tooltips}{tooltip (html is accepted) text.}
}
\description{
add tooltips on graphical elements.
}

