% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{set_data_id}
\alias{set_data_id}
\title{add id}
\usage{
set_data_id(ids, datid)
}
\arguments{
\item{ids}{integer vector of graphical elements identifiers (returned by
\code{\link{rvg_tracer_off}}).}

\item{datid}{user id to be associated with \code{ids}.}
}
\description{
add id to graphical elements.
}

