% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Microsoft Excel Graphics Device}
\usage{
write_xlsx(file, code, size = c(width = 10, height = 7.5), ...)
}
\arguments{
\item{file}{filename of the Microsoft Excel document to produce. File
extension must be \code{.xlsx}.}

\item{code}{Plotting code to execute}

\item{size}{slide size in inches.}

\item{...}{arguments for \code{fun} (passed on to \code{\link{dml_xlsx}}.)}
}
\description{
A graphical device for Microsoft Excel documents.
}
\examples{
write_xlsx(file = "my_plot_1.xlsx", code = plot(rnorm(10)))
write_xlsx(file = "my_plot_2.xlsx", code = barplot(1:7, col = 1:7))
}
\keyword{device}

